PyShare
Updated version of file share. The code is more module, spelling has been
fixed and silly mistakes have been fixed. The program is now much better
and since it is not and of the same code from before It is going to be a
new project completely. The program is much more stable.